requirements.txt
langgraph==0.2.28
langchain==0.3.7
langchain-groq==0.2.1
groq==0.11.0
pinecone-client==5.0.1
streamlit>=1.40.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
sentence-transformers==3.2.1
pandas==2.2.3
requests==2.32.3
pydantic==2.9.2
pillow>=11.0.0